requirements.txt
flask==2.0.1
tensorflow>=2.6.0
torch>=1.10.0
torchvision>=0.11.1
numpy>=1.19.5
matplotlib>=3.4.3
pillow>=8.3.2
reportlab>=3.6.1
fpdf2>=2.5.1
fpdf==1.7.2
pdfkit>=1.0.0
opencv-python>=4.5.3
ngrok>=1.4.0
